rename klocal to local
