+++ provider/local/template/config.yml.tmpl
@@ -8,4 +8,4 @@ metadata:
     type: config
     rack: {{.Rack}}
 data:
-  PROVIDER: klocal
+  PROVIDER: local
