+++ provider/local/template/resource/memcached.yml.tmpl
