+++ provider/local/template/resource/mysql.yml.tmpl
