+++ provider/local/template/resource/redis.yml.tmpl
